Caching and loading improvements
