+++ ui/.env
@@ -1,3 +1,3 @@
 BACKEND_URL=https://console-api-eyefloaters-dev.mycluster-us-east-107719-da779aef12eee96bd4161f4e402b70ec-0000.us-east.containers.appdomain.cloud
 LOG_LEVEL="info"
-CONSOLE_MODE="read-write"
+CONSOLE_MODE="read-only"
