+++ ui/api/resources.ts
@@ -55,7 +55,6 @@ export async function getResources(
   scope: typeof ResourceTypeKafka,
 ): Promise<KafkaResource[]>;
 export async function getResources(scope: unknown): Promise<unknown> {
-  "use server";
   const { resources } = await getSession("resources");
   switch (scope) {
     case "kafka":
@@ -91,7 +90,6 @@ export async function getResource(
   id: string,
   scope: unknown,
 ): Promise<unknown> {
-  "use server";
   let resource;
   if (typeof scope === typeof ResourceTypeKafka) {
     const resources = await getResources("kafka");
