+++ ui/app/[locale]/kafka/[kafkaId]/@modal/topics/[topicId]/delete/page.tsx
@@ -1,0 +1,3 @@
+export default function DeletePage() {
+  return null;
+}
