+++ ui/app/[locale]/kafka/[kafkaId]/topics/page.tsx
@@ -1,38 +1,50 @@
-import { getTopics, TopicList } from "@/api/topics";
-import { TopicsTable } from "@/app/[locale]/kafka/[kafkaId]/topics/TopicsTable";
+import { getTopics } from "@/api/topics";
+import {
+  TopicsTable,
+  TopicsTableColumn,
+} from "@/app/[locale]/kafka/[kafkaId]/topics/TopicsTable";
 import { PageSection } from "@/libs/patternfly/react-core";
+import { stringToInt } from "@/utils/stringToInt";
+
+const sortMap: Record<TopicsTableColumn, string> = {
+  name: "name",
+  messages: "recordCount",
+  partitions: "partitions",
+  storage: "totalLeaderLogBytes",
+};
 
 export default async function TopicsPage({
   params,
+  searchParams,
 }: {
   params: { kafkaId: string };
   searchParams: {
-    limit: string | undefined;
+    perPage: string | undefined;
     sort: string | undefined;
-    "filter[offset]": string | undefined;
-    "filter[timestamp]": string | undefined;
-    "filter[epoch]": string | undefined;
+    sortDir: string | undefined;
+    page: string | undefined;
   };
 }) {
-  const topics = await getTopics(params.kafkaId);
-  return (
-    <TopicsContent
-      canCreate={process.env.CONSOLE_MODE === "read-write"}
-      topics={topics}
-    />
-  );
-}
-
-function TopicsContent({
-  topics,
-  canCreate,
-}: {
-  topics: TopicList[];
-  canCreate: boolean;
-}) {
+  const pageSize = stringToInt(searchParams.perPage) || 10;
+  const sort = (searchParams["sort"] || "name") as TopicsTableColumn;
+  const sortDir = (searchParams["sortDir"] || "asc") as "asc" | "desc";
+  const pageCursor = searchParams["page"];
+  const topics = await getTopics(params.kafkaId, {
+    sort: sortMap[sort],
+    sortDir,
+    pageSize,
+    pageCursor,
+  });
   return (
     <PageSection isFilled>
-      <TopicsTable topics={topics} canCreate={canCreate} />
+      <TopicsTable
+        topics={topics.data}
+        topicsCount={topics.meta.page.total}
+        perPage={pageSize}
+        sort={sort}
+        sortDir={sortDir}
+        canCreate={process.env.CONSOLE_MODE === "read-write"}
+      />
     </PageSection>
   );
 }
